# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-grade Go backend for Hackathon Intelligence.",
        "title": "Hackathon Intelligence Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://hackbuddy-im66.onrender.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{id}/analyses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all analyses run for the session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "List analyses for a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analyses array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/analysis/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns latest analysis, default directive, and custom directives for the session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Get analysis summary with directive for a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AnalysisSummary in data (latest_analysis, default_directive, custom_directives)",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/analyze": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Analyzes session sources and stores structured results plus a default directive. Optional body: directives (strings) to steer the analysis and assistant.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Run analysis on a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional directives",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domains_analysis.AnalyzeInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Analysis in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Analysis failed",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{id}/chat": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all chat messages (user and assistant) for the session, ordered by creation time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat history for a session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a message in a session. Uses session context and directive. Returns assistant reply.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Send a chat message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message content",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains_chat.ChatInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assistant message in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Validation or guardrail",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domains_analysis.AnalyzeInput": {
            "type": "object",
            "properties": {
                "directives": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "domains_chat.ChatInput": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "hackbuddy-backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "{} | [] | null"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" | \"error\"",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}